system/parse.go: refactor, merge with system.go
